refactor: mover CredenciaisUsuario para seu próprio arquivo
